moves Functions into appropriate modules
