+++ lib/PureScript/Ide/Types.hs
@@ -50,4 +50,5 @@ instance ToJSON Completion where
   toJSON (Completion (m, d, t)) =
     object ["module" .= m, "identifier" .= d, "type" .= t]
 
+type CompletionFilter = [Module] -> [Module]
 type Matcher = [Completion] -> [Completion]
